Add Cmake script
