+++ include/encoding.h
